+++ libsolidity/parsing/Parser.h
@@ -171,6 +171,8 @@ private:
 	FunctionCallArguments parseFunctionCallArguments();
 	FunctionCallArguments parseNamedArguments();
 	std::pair<ASTPointer<ASTString>, langutil::SourceLocation> expectIdentifierWithLocation();
+
+	ASTPointer<Expression> parseContractStorageBaseLocationExpression();
 	///@}
 
 	///@{