+++ test/libsolidity/syntaxTests/contractBaseLocation/layout_specification_constant_in_expression.sol
@@ -1,0 +1,3 @@
+uint constant X = 42;
+contract C layout at 0xffff * (50 - X) { }
+// ----