+++ test/libsolidity/syntaxTests/contractBaseLocation/layout_specification_no_expression.sol
@@ -1,0 +1,3 @@
+contract C layout at { }
+// ----
+// ParserError 6933: (21-22): Expected primary expression.